Cambio al tipo de datos
Se cambio el tipo de dato de double a float, para
que sea acorde con el resto del programa, pues
la calculadora recibe en el mapa valores de tipo
float

+++ variable_user_input.h
@@ -1,9 +1,9 @@
 #ifndef VARIABLE_USER_INPUT
 #define VARIABLE_USER_INPUT
 
-double get_variable_input(string variable) {
+float get_variable_input(string variable) {
 
-    double input_var;
+    float input_var;
 
     std::cout << "Enter value for (" <<variable << "): ";
     std::cin >> input_var;
